refactor(types): restructure NavItem type and update its imports
