+++ src/types/index.ts
@@ -5,3 +5,4 @@ export * from './Project';
 export * from './SidebarItem';
 export * from './SortOption';
 export * from './Tool';
+export * from './NavItem';
